# Use WithLabelValues for metrics in hit_tracker_client