I also got rid of a proto clone which doesn't seem to be necessary. It's only used when flushing synchronously.

This should save 0.2% of CPU in the cache proxy.

Benchmark results:
```
goos: linux
goarch: amd64
cpu: AMD Ryzen 9 7900X 12-Core Processor
           │     base     │               values               │           values-noclone           │
           │    sec/op    │   sec/op     vs base               │   sec/op     vs base               │
Enqueue-24   530.0m ± 10%   512.0m ± 3%  -3.38% (p=0.010 n=11)   493.1m ± 5%  -6.95% (p=0.000 n=11)

           │     base     │                values                │            values-noclone             │
           │     B/op     │     B/op      vs base                │     B/op       vs base                │
Enqueue-24   1.607Gi ± 7%   1.318Gi ± 9%  -17.98% (p=0.000 n=11)   1.247Gi ± 10%  -22.41% (p=0.000 n=11)

           │    base     │               values               │           values-noclone            │
           │  allocs/op  │  allocs/op   vs base               │  allocs/op   vs base                │
Enqueue-24   21.43M ± 5%   19.44M ± 5%  -9.29% (p=0.000 n=11)   19.27M ± 6%  -10.08% (p=0.000 n=11)
```