@iain-macdonald 2025-06-04
> The proto.Clone thing you noticed is actually a bug, we should be sending the RequestMetadata along with asynchronous hits too.

Oh, we already do that in `CloseWithBytesTransferred`! I think you can just get rid of that function parameter entirely.